+++ src/link/persisted-queries/index.ts
@@ -10,7 +10,12 @@ import {
 import { invariant } from 'ts-invariant';
 
 import { ApolloLink, Operation } from '../core';
-import { Observable, Observer, compact } from '../../utilities';
+import {
+  Observable,
+  Observer,
+  ObservableSubscription,
+  compact,
+} from '../../utilities';
 
 export const VERSION = 1;
 
@@ -147,7 +152,7 @@ export const createPersistedQueryLink = (
     const { query } = operation;
 
     return new Observable((observer: Observer<ExecutionResult>) => {
-      let subscription: ZenObservable.Subscription;
+      let subscription: ObservableSubscription;
       let retried = false;
       let originalFetchOptions: any;
       let setFetchOptions = false;